--- alembic/versions/8b37089f95cb_app_object.py
@@ -1,42 +1,0 @@
-"""App object
-
-Revision ID: 8b37089f95cb
-Revises: 5c0b6ae8f9ac
-Create Date: 2023-10-18 18:26:08.188463
-
-"""
-from typing import Sequence, Union
-
-from alembic import op
-import sqlalchemy as sa
-
-
-# revision identifiers, used by Alembic.
-revision: str = '8b37089f95cb'
-down_revision: Union[str, None] = '5c0b6ae8f9ac'
-branch_labels: Union[str, Sequence[str], None] = None
-depends_on: Union[str, Sequence[str], None] = None
-
-
-def upgrade() -> None:
-    # ### commands auto generated by Alembic - please adjust! ###
-    op.create_table('apps',
-    sa.Column('id', sa.Integer(), nullable=False),
-    sa.Column('user_ref', sa.Integer(), nullable=False),
-    sa.Column('app_name', sa.String(length=100), nullable=False),
-    sa.Column('app_description', sa.Text(), nullable=False),
-    sa.Column('app_key', sa.String(length=50), nullable=False),
-    sa.Column('app_model', sa.String(length=50), nullable=False),
-    sa.Column('app_temperature', sa.Float(), nullable=False),
-    sa.ForeignKeyConstraint(['user_ref'], ['users.id'], onupdate='CASCADE', ondelete='CASCADE'),
-    sa.PrimaryKeyConstraint('id', 'user_ref'),
-    sa.UniqueConstraint('user_ref', 'app_key', name='_app_key_unique'),
-    sa.UniqueConstraint('user_ref', 'app_name', name='_app_name_unique')
-    )
-    # ### end Alembic commands ###
-
-
-def downgrade() -> None:
-    # ### commands auto generated by Alembic - please adjust! ###
-    op.drop_table('apps')
-    # ### end Alembic commands ###
