+++ alembic/versions/9c116104177d_initial_migration.py
@@ -1,0 +1,65 @@
+"""initial migration
+
+Revision ID: 9c116104177d
+Revises: 
+Create Date: 2023-10-19 22:52:09.210053
+
+"""
+from typing import Sequence, Union
+
+from alembic import op
+import sqlalchemy as sa
+
+
+# revision identifiers, used by Alembic.
+revision: str = '9c116104177d'
+down_revision: Union[str, None] = None
+branch_labels: Union[str, Sequence[str], None] = None
+depends_on: Union[str, Sequence[str], None] = None
+
+
+def upgrade() -> None:
+    # ### commands auto generated by Alembic - please adjust! ###
+    op.create_table('users',
+    sa.Column('id', sa.Integer(), autoincrement=True, nullable=False),
+    sa.Column('email', sa.String(length=250), nullable=False),
+    sa.Column('password', sa.String(length=200), nullable=False),
+    sa.Column('name', sa.String(length=50), nullable=False),
+    sa.PrimaryKeyConstraint('id'),
+    sa.UniqueConstraint('email')
+    )
+    op.create_table('apps',
+    sa.Column('id', sa.Integer(), autoincrement=True, nullable=False),
+    sa.Column('user_ref', sa.Integer(), nullable=False),
+    sa.Column('app_name', sa.String(length=100), nullable=False),
+    sa.Column('app_description', sa.Text(), nullable=False),
+    sa.Column('app_key', sa.String(length=50), nullable=False),
+    sa.Column('app_model', sa.String(length=50), nullable=False),
+    sa.Column('app_temperature', sa.Float(), nullable=False),
+    sa.ForeignKeyConstraint(['user_ref'], ['users.id'], onupdate='CASCADE', ondelete='CASCADE'),
+    sa.PrimaryKeyConstraint('id'),
+    sa.UniqueConstraint('user_ref', 'app_key', name='_app_key_unique'),
+    sa.UniqueConstraint('user_ref', 'app_name', name='_app_name_unique')
+    )
+    op.create_table('chat_messages',
+    sa.Column('id', sa.Integer(), autoincrement=True, nullable=False),
+    sa.Column('user_ref', sa.Integer(), nullable=False),
+    sa.Column('chatbot_id', sa.String(length=50), nullable=False),
+    sa.Column('message', sa.Text(), nullable=False),
+    sa.Column('is_bot_reply', sa.Boolean(), nullable=False),
+    sa.Column('createdat', sa.DateTime(), server_default=sa.text('now()'), nullable=False),
+    sa.ForeignKeyConstraint(['user_ref'], ['users.id'], onupdate='CASCADE', ondelete='CASCADE'),
+    sa.PrimaryKeyConstraint('id'),
+    sa.UniqueConstraint('chatbot_id')
+    )
+    op.create_index(op.f('ix_chat_messages_createdat'), 'chat_messages', ['createdat'], unique=False)
+    # ### end Alembic commands ###
+
+
+def downgrade() -> None:
+    # ### commands auto generated by Alembic - please adjust! ###
+    op.drop_index(op.f('ix_chat_messages_createdat'), table_name='chat_messages')
+    op.drop_table('chat_messages')
+    op.drop_table('apps')
+    op.drop_table('users')
+    # ### end Alembic commands ###
